fix: (widen XP accumulation to Long and cap rule config so it cannot overflow)
RuleField validated a floor but no ceiling, so an absurd xpPerMinute was
accepted and Gamification.gameXp wrapped it as Int; levelState's
coerceAtLeast(0) then turned the wrapped negative into a plausible-looking
totalXp = 0, level = 1. The same unchecked Int accumulation could also
overflow across a large-enough library or achievement count with no
absurd setting at all (#114).

RuleField now rejects a value above a documented per-field ceiling with
headroom under the widened bound, and Gamification/XpState accumulate in
Long throughout, so no accepted configuration can wrap the total. Every
consumer of totalXp/XpState across the app widened to match; single-value
fields that are provably bounded by a per-tier ceiling (one achievement's
XP) stay Int at their display boundary.

PlayerProfile.totalXp widens to Long (v28->v29, additive only — SQLite's
INTEGER affinity already covers 64 bits, so no ALTER is needed for the
column itself). Every profile whose totalXp reads 0 at migration time is
flagged pendingXpIntegrityCorrection, since that shape is indistinguishable
from an overflow victim; the next recompute declares
RecomputeSource.XP_INTEGRITY_CORRECTION regardless of its caller's own
source, so a large upward correction reseeds the delivery baseline instead
of firing a cascade of level-up events for progress earned long ago —
the same defect #104 is about, avoided here while fixing #114.

+++ app/src/main/java/com/example/backlogium/data/repo/ProfileRepository.kt
@@ -15,7 +15,7 @@ import javax.inject.Singleton
 data class PlayerStats(
     val steamId: String,
     val steamLevel: Int,
-    val totalXp: Int,
+    val totalXp: Long,
     val level: Int,
     val currentStreak: Int,
     val longestStreak: Int,
